# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Open Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get list of pages",
                "operationId": "9e05f7ba039b9d59de1e43b0a78ca181",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/query_limit"
                    },
                    {
                        "$ref": "#/components/parameters/query_page"
                    },
                    {
                        "$ref": "#/components/parameters/query_keyword"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PageResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Create a new page",
                "operationId": "d810dbe9fe40d6fcc6014a7320c6d99b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/pages/{id}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get page details by id",
                "operationId": "577c2ec499b66899010c542d14d3194c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PageResource"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update an existing page",
                "operationId": "ebe61096ddf4608a7e35a512582c8b6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Delete a page",
                "operationId": "6933f58f61835a011bb8ab33a9f51c7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/pages/bulk": {
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Bulk action on pages",
                "operationId": "8b3a1acaa630c61ce72bc51650fd7dde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "action": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bulk action successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PageRequest": {
                "required": [
                    "title",
                    "locale",
                    "status"
                ],
                "properties": {
                    "locale": {
                        "type": "string",
                        "example": "en"
                    },
                    "status": {
                        "type": "string",
                        "example": "published"
                    },
                    "title": {
                        "type": "string",
                        "example": "Page Title"
                    },
                    "slug": {
                        "type": "string",
                        "example": "page-slug"
                    },
                    "content": {
                        "type": "string",
                        "example": "Page content"
                    },
                    "template": {
                        "type": "string",
                        "example": "default"
                    },
                    "thumbnail": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PageResource": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string",
                        "example": "About Us"
                    },
                    "slug": {
                        "type": "string",
                        "example": "about-us"
                    },
                    "description": {
                        "type": "string",
                        "example": "About us description"
                    },
                    "content": {
                        "type": "string",
                        "example": "About us content"
                    },
                    "template": {
                        "type": "string",
                        "example": "page"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "total_pages": {
                        "type": "integer"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "page": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "self": {
                        "type": "string"
                    },
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "type": "string"
                    },
                    "next": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
...
            "query_limit": {
                "name": "limit",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "query_page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "query_keyword": {
                "name": "keyword",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "in": "header",
                "name": "Authorization"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}